Reference

Reference

Agent generation

SmartTransitionSim.Rect — Type.

Rect type represent rectangle area on map

Fields

p1 : Latitude-Longitude coordinates of first rectangle corner

p2 : Latitude-Longitude coordinates of second rectangle corner

source

SmartTransitionSim.Agent — Type.

Agent type stores information about agents

Fields

smart : logical value indicating if agent use re-routing service

start_node : starting point of agent's route

end_node : ending point of agent's route

route : array of nodes determining agent's route (may be changed by re-routing)

travel_time : time spent in simulation

edge : current edge agent is on

active : indicates if agent is active in simulation

source

SmartTransitionSim.pick_random_node — Function.

pick_random_node function is used to set starting and ending node of agents. Nodes are randomly chosen from set of provided areas.

Input parameters

OSMmap : OpenStreetMapX mapData object with road network data

rects : vector of Rect types interpreted as a set of rectangle areas

NodesSet : returning object switch

false return random node from given area

true return all nodes in given area for further processing

source

SmartTransitionSim.generate_agents — Function.

generate_agents function creating vector of agents and returning travel time for initial routes travelled with maximal speed

Input parameters

OSMmap : OpenStreetMapX MapData object with road network data

N : number of agents to be generated

StartArea : vector of areas from which agents randomly pick starting point

EndArea : vector of areas from which agents randomly pick ending point

α : percentage of smart agents

source

Rerouting

SmartTransitionSim.k_shortest_path_rerouting! — Function.

k_shortest_path_rerouting! change agents current route according to traffic conditions

Input parameters

OSMmap : OpenStreetMapX MapData object with road network data

inAgent : modified agent

speeds : current speed matrix

k : number of fastest routes returned

T : control variable for probability distribution

source

Traffic model

Missing docstring.

Missing docstring for get_max_densities. Check Documenter's build log for details.

SmartTransitionSim.update_weights! — Function.

update_weights! change speeds in given speedmatrix for edges listed in newdensities

Input parameters

speed_matrix : matrix with average speeds on edges

new_densities : dictionary with edges as keys and new traffic density as value

V_max : matrix with maximal speeds on edges

V_min : minimal speed on road

source

SmartTransitionSim.traffic_constants — Function.

traffic_constants create maximal traffic densities and speeds matrices

Input parameters

OSMmap : OpenStreetMapX MapData object with road network data

density_factor : road length reserved for one vehicle

source

SmartTransitionSim.init_traffic_variables — Function.

init_traffic_variables create data used in calculating velocities change during simulation

Input parameters

OSMmap : OpenStreetMapX MapData object with road network data

Agents : set of agents created with generate_agents function

source

SmartTransitionSim.next_edge — Function.

next_edge returns time required to reach next junction for one agent

Input parameters

Agent : single instance of Agent

speeds : current speeds matrix

lengths : road lengths matrix

source

next_edge returns time required to reach next junction for set of agents

Input parameters

Agents : set of agents created with generate_agents function

speeds : current speeds matrix

lengths : road lengths matrix

source

SmartTransitionSim.update_event_agent! — Function.

update_event_agent! update densities matrix, progress agents to next edge and deactivate agents

Input parameters

inAgent : agent connected with occuring event

curr_time : current simulation time

densities : current traffic densitites dictionary

vertices_map : mapping from nodes to vertices

source

Missing docstring.

Missing docstring for update_agents_position!. Check Documenter's build log for details.